+++ NeoMathEngine/src/CPU/x86/avx/src/JitCommon.h
@@ -186,11 +186,11 @@ public:
 
 private:
     struct CLoopDesc {
-        CLoopDesc( reg64_t counter, size_t step ) : Counter( counter ), Step( step ) {}
+        CLoopDesc( reg64_t counter, uint32_t step ) : Counter( counter ), Step( step ) {}
         Xbyak::Label StartLabel;
         Xbyak::Label EndLabel;
         reg64_t Counter;
-        size_t Step;
+        uint32_t Step;
     };
     std::stack<CLoopDesc> loopDescs;
